Add CLI to main and move models to agtern/utils
