--- agtern.py
@@ -1,4 +1,0 @@
-if __name__ == "__main__":
-    from agtern.utils import run_cli
-
-    run_cli()
