+++ agtern/__init__.py
@@ -1,4 +1,3 @@
-from .__main__ import HEADLESS_SCRAPER, main
 from .gui import *
 from .pipelines import *
 from .utils import *
