+++ agtern/__main__.py
@@ -1,5 +1,7 @@
-from agtern.gui import Application
-from agtern.pipelines import start_server
+from argparse import ArgumentParser
+
+from .gui import Application
+from .pipelines import import_companies, sort_companies, start_server
 
 HEADLESS_SCRAPER = True
 
@@ -18,5 +20,21 @@ def main(noscrape: bool = True, headless_scraper=None):
         pass  # Do nothing and hide Traceback
 
 
+def run_cli(args=None):
+    parser = ArgumentParser(prog="AgTern")
+    parser.add_argument("--update-companies", action="store_true")
+    parser.add_argument("--show-scraper", action="store_true")
+    parser.add_argument("--noscrape", action="store_true")
+    args = parser.parse_args(args)
+
+    if args.update_companies:
+        print("INFO: Updating company info...")
+        sort_companies()
+        import_companies()
+    else:
+        print("INFO: Starting program...")
+        main(noscrape=args.noscrape, headless_scraper=not args.show_scraper)
+
+
 if __name__ == "__main__":
-    main()
+    run_cli()
