+++ agtern/pipelines/utils/__init__.py
@@ -1,2 +1,1 @@
 from .data import *
-from .models import *
