+++ agtern/pipelines/utils/data/data_file.py
@@ -17,25 +17,25 @@ class DataFile:
             self.path = Path.cwd().joinpath("data", "temp", filename)
         else:
             self.path = Path.cwd().joinpath("data", filename)
-
-        # Makes sure file exists after initialization
-        assert self.exists() or self.create(default_data)
+        if default_data is not None:
+            self.create(default_data)
 
     def exists(self):
         return isfile(self.path)
 
-    def create(self, default_data: str):
+    def create(self, default_data: Union[str, None] = None):
         """Creates a data file if it doesn't exist. Returns true if successful."""
-        # Will only create new file if default data given
-        assert not self.exists() and default_data
-        try:
-            if self.is_temp:
-                makedirs(Path.cwd().joinpath("data", "temp"), exist_ok=True)
-            else:
-                makedirs(Path.cwd().joinpath("data"), exist_ok=True)
-
-            with open(self.path, "w") as f:
-                f.write(default_data)
-        except OSError:
-            pass
+        if self.exists():
+            return True
+        if default_data is not None:
+            try:
+                if self.is_temp:
+                    makedirs(Path.cwd().joinpath("data", "temp"), exist_ok=True)
+                else:
+                    makedirs(Path.cwd().joinpath("data"), exist_ok=True)
+                with open(self.path, "w") as f:
+                    if default_data is not None:
+                        f.write(default_data)
+            except OSError:
+                pass
         return self.exists()
