+++ agtern/utils/__init__.py
@@ -1,1 +1,1 @@
-from .cli import *
+from .models import *
