--- agtern/utils/cli/__init__.py
@@ -1,1 +1,0 @@
-from .run_cli import run_cli
